# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api-IvoireCouture",
        "version": "1.0"
    },
    "paths": {
        "/api/clients": {
            "get": {
                "tags": [
                    "clients"
                ],
                "summary": "list des clients",
                "description": "Obtenir la liste de tous les clients",
                "operationId": "client_list",
                "responses": {
                    "200": {
                        "description": "Afficher la liste de tous les clients"
                    }
                }
            },
            "post": {
                "tags": [
                    "clients"
                ],
                "summary": "Créer un client",
                "description": "Création d'un client",
                "operationId": "clients",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nom": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "sexe": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "nom": "Kouadio",
                                    "phone": "0102030405",
                                    "image": "lien de l'image",
                                    "sexe": "h"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "405": {
                        "description": "Exception"
                    }
                }
            }
        },
        "/api/clients/{id}": {
            "get": {
                "tags": [
                    "clients"
                ],
                "summary": "Obtenir un seul client",
                "description": "Obtenir un seul client",
                "operationId": "one_client",
                "responses": {
                    "200": {
                        "description": "Afficher un seul client"
                    },
                    "404": {
                        "description": "Utilisateur intouvable ou inexistant"
                    }
                }
            },
            "put": {
                "tags": [
                    "clients"
                ],
                "summary": "Mise à jour d'un client",
                "description": "Mise à jour des données d'un client",
                "operationId": "update_clients",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nom": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "sexe": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "nom": "Kouadio",
                                    "phone": "0102030405",
                                    "image": "lien modifié",
                                    "sexe": "f"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "405": {
                        "description": "Exception"
                    }
                }
            },
            "delete": {
                "tags": [
                    "clients"
                ],
                "summary": "Supprimer un client",
                "description": "Supprimer un client",
                "operationId": "delete_client",
                "responses": {
                    "200": {
                        "description": "Supprime un seul client"
                    },
                    "404": {
                        "description": "client intouvable ou inexistant"
                    }
                }
            }
        },
...
    }
}